requirements.txt
alembic==0.9.7
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.10.1
Flask-Bootstrap==3.3.7.1
Flask-CLI==0.4.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==19.4.5
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.1
pycountry==18.2.23
PyJWT==1.6.1
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
